Document Gemfile
To make it easier to see what gems are for, I have documents gems where the names are not clear of what they are used for.

I have also grouped them by use to make it easier to see similar gems.

Also discovered on newer ruby versions posix-spawn is not required.